requirements.txt
pandas
requests
pytest
fastapi
uvicorn
pyiceberg
python-dotenv
python-multipart
langchain
langchain-google-genai
langchain-core
langgraph
sqlalchemy
passlib
bcrypt==4.0.1
python-jose[cryptography]
email-validator
tenacity
pydantic
supabase==2.0.2
celery
redis
nest_asyncio
python-socketio
aiohttp
google-generativeai
httpx
httpcore
reportlab
openpyxl
sentry-sdk[fastapi]
